server/requirements.txt
Flask==1.1.2
Flask-API==2.0
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-MySQL==1.5.2
flask-mysql-connector==1.0.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.1
mysql-connector==2.2.9
mysql-connector-python==8.0.23
mysqlclient==2.0.3
PyMySQL==1.0.2
requests==2.25.1
SQLAlchemy==1.3.18
sqlparse==0.4.1